pyproject.toml
[tool.poetry]
name = "get_papers"
version = "0.1.0"
description = "CLI tool to fetch PubMed papers with biotech/pharma authors"
authors = ["Pratiksha Kutal"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
lxml = "^5.3.1"
click = "^8.1.8"
biopython = "^1.85"


[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "get_papers_cli:main"
